+++ NEWS.md
@@ -1,3 +1,17 @@
+
+
+## SmallCountRounding	0.7.0
+* Now, a random generator seed is used locally within the function without affecting the random value stream in R.
+  - Handled by `rndSeed`, a new parameter to `RoundViaDummy` (and `PLSrounding`). 
+  - By default, `rndSeed = 123`. This means that repeated runs with equal input will result in equal output. 
+  - To get back the old behaviour of the function, set `rndSeed` to `NULL`.
+* Possible to return a single data frame: `"inner"` or `"publish"`.
+  - Handled by `output`, a new parameter to `PLSrounding`. 
+  - New wrapper functions, `PLSroundingInner` and  `PLSroundingPublish`.
+* `dimVar`	is new parameter to `RoundViaDummy` and `PLSrounding`
+  - The main dimensional variables and additional aggregating variables. This parameter can be  useful when hierarchies and formula are unspecified. 
+
+
 ## SmallCountRounding	0.6.0
 * `preRounded`	is new parameter to `RoundViaDummy` (and `PLSrounding`)
   - A mixture of missing values and predetermined values of rounded inner cells